Restructure package under app/ and clean old scripts
