+++ start.py
@@ -37,8 +37,8 @@ def main():
         
         # Start the FastAPI server
         subprocess.run([
-            sys.executable, "-m", "uvicorn", 
-            "main:app", 
+            sys.executable, "-m", "uvicorn",
+            "app.main:app",
             "--reload", 
             "--host", "0.0.0.0", 
             "--port", "8000"
